+++ gemrb/plugins/TLKImporter/TLKImp.h
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  *
- * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/TLKImporter/TLKImp.h,v 1.16 2004/11/28 19:36:39 avenger_teambg Exp $
+ * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/TLKImporter/TLKImp.h,v 1.17 2005/02/23 20:59:38 guidoj Exp $
  *
  */
 
@@ -35,8 +35,8 @@ public:
 	TLKImp(void);
 	~TLKImp(void);
 	bool Open(DataStream* stream, bool autoFree = true);
-	char* GetString(ieStrRef strref, unsigned long flags = 0);
-	StringBlock GetStringBlock(ieStrRef strref, unsigned long flags = 0);
+	char* GetString(ieStrRef strref, unsigned int flags = 0);
+	StringBlock GetStringBlock(ieStrRef strref, unsigned int flags = 0);
 private:
 	/** resolves day and monthname tokens */
 	void GetMonthName(int dayandmonth);
